done Ultrasonic

+++ Ultrasonic/Inc/median.h
@@ -1,25 +1,18 @@
-/*
- * median.h
- *
- * Created: 18.10.2016 15:25:10
- *  Author: M43888
- */
-
 #ifndef MEDIAN_H_
 #define MEDIAN_H_
 
 #include <stdint.h>
 
 typedef struct _median_filter {
-	float  length;
-	float  median;
+	uint16_t  length;
+	uint16_t  median;
 	uint8_t   counter;
-	float *value_bfr;
-	float *sort_bfr;
+	uint16_t *value_bfr;
+	uint16_t *sort_bfr;
 } median_filter_t;
 
-void median_filter_init(median_filter_t *filter, float lenght, float *bfr, float *sort_bfr);
+void median_filter_init(median_filter_t *filter, uint16_t lenght, uint16_t *bfr, uint16_t *sort_bfr);
 
-float median_filter_add_new_value(median_filter_t *filter, float new_value);
+uint16_t median_filter_add_new_value(median_filter_t *filter, uint16_t new_value);
 
 #endif /* MEDIAN_H_ */
